requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dash==1.9.1
dash-bootstrap-components==0.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
plotly==4.5.4
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-cache==0.5.2
retrying==1.3.3
Shapely==1.7.0
six==1.14.0
topojson==1.0rc8
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
Werkzeug==1.0.0
